# apps/rust-server/openapi.generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "InstantML Rust API",
    "description": "Current Rust/ClickHouse API. See docs/architecture/current-api.md for examples and operational notes.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/invitations/accept": {
      "post": {
        "tags": [
          "invitations"
        ],
        "operationId": "accept_invitation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session payload for accepted organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionPayload"
                }
              }
            }
          },
          "401": {
            "description": "Browser session required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Invite email mismatch or fresh Clerk exchange required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "Invitation expired or revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "browserSession": []
          }
        ]
      }
    },
    "/api/invitations/preview": {
      "post": {
        "tags": [
          "invitations"
        ],
        "operationId": "preview_invitation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Safe invitation preview",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationPreviewEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Invitation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Invitation preview rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/api/orgs/{org_id}/invitations": {
      "get": {
        "tags": [
          "orgs"
        ],
        "operationId": "list_invitations",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "description": "Organization UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum rows to return; defaults to 100",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Rows to skip",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Organization invitations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationsEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "browserSession": []
          },
          {
            "bootstrapToken": []
          }
        ]
      },
      "post": {
        "tags": [
          "orgs"
        ],
        "operationId": "create_invitation",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "description": "Organization UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Created organization invitation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "browserSession": []
          },
          {
            "bootstrapToken": []
          }
        ]
      }
    },
    "/api/orgs/{org_id}/invitations/{invitation_id}/resend": {
      "post": {
        "tags": [
          "orgs"
        ],
        "operationId": "resend_invitation",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "description": "Organization UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invitation_id",
            "in": "path",
            "description": "Invitation UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resent organization invitation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "browserSession": []
          },
          {
            "bootstrapToken": []
          }
        ]
      }
    },
    "/api/orgs/{org_id}/invitations/{invitation_id}/revoke": {
      "post": {
        "tags": [
          "orgs"
        ],
        "operationId": "revoke_invitation",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "description": "Organization UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invitation_id",
            "in": "path",
            "description": "Invitation UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Revoked organization invitation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "browserSession": []
          },
          {
            "bootstrapToken": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthSessionPayload": {
        "type": "object",
        "required": [
          "authenticated",
          "session",
          "user",
          "organization",
          "membership",
          "memberships",
          "account_type"
        ],
        "properties": {
          "account_type": {
            "type": "string"
          },
          "authenticated": {
            "type": "boolean"
          },
          "billing_checkout": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BillingCheckoutInfo"
              }
            ]
          },
          "membership": {
            "$ref": "#/components/schemas/MembershipRow"
          },
          "memberships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MembershipRow"
            }
          },
          "organization": {
            "$ref": "#/components/schemas/OrganizationRow"
          },
          "provisioning": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ProvisioningStatusPayload"
              }
            ]
          },
          "session": {
            "$ref": "#/components/schemas/UserSessionRow"
          },
          "user": {
            "$ref": "#/components/schemas/UserRow"
          }
        }
      },
...
      "BillingCheckoutInfo": {
        "type": "object",
        "required": [
          "intent_id",
          "status"
        ],
        "properties": {
          "intent_id": {
            "type": "string",
            "format": "uuid"
          },
          "session_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "url": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "CreateInvitationRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "role": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "InvitationEnvelope": {
        "type": "object",
        "required": [
          "invitation"
        ],
        "properties": {
          "delivery_error": {
            "type": [
              "string",
              "null"
            ]
          },
          "invitation": {
            "$ref": "#/components/schemas/PublicInvitationRow"
          },
          "preview_link": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "InvitationPreviewEnvelope": {
        "type": "object",
        "required": [
          "invitation"
        ],
        "properties": {
          "invitation": {
            "$ref": "#/components/schemas/InvitationPreviewPayload"
          }
        }
      },
      "InvitationPreviewPayload": {
        "type": "object",
        "required": [
          "org_name",
          "email_hint",
          "role",
          "status",
          "expires_at"
        ],
        "properties": {
          "email_hint": {
            "type": "string"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "org_name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "InvitationTokenRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "InvitationsEnvelope": {
        "type": "object",
        "required": [
          "invitations",
          "limit",
          "offset"
        ],
        "properties": {
          "invitations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicInvitationRow"
            }
          },
          "limit": {
            "type": "integer",
            "minimum": 0
          },
          "offset": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "MembershipRow": {
        "type": "object",
        "required": [
          "id",
          "org_id",
          "user_id",
          "role",
          "status",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "org_id": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "OrganizationRow": {
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "plan_tier",
          "account_type",
          "seat_limit",
          "created_at"
        ],
        "properties": {
          "account_type": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_by_user_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "plan_tier": {
            "type": "string"
          },
          "seat_limit": {
            "type": "integer",
            "format": "int32"
          },
          "slug": {
            "type": "string"
          },
          "tenant_routing_tier": {
            "type": "string",
            "description": "Routing tier for this org's ClickHouse data plane.\n`\"shared\"` — routes to the shared cell (free/personal orgs).\n`\"dedicated\"` — routes to a per-org provisioned service.\nOlder records that pre-date this field deserialize to `\"dedicated\"`\n(the safe fallback) via the serde default."
          }
        }
      },
...
      "ProvisioningStatusPayload": {
        "type": "object",
        "required": [
          "status",
          "mode"
        ],
        "properties": {
          "mode": {
            "type": "string"
          },
          "service_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "PublicInvitationRow": {
        "type": "object",
        "required": [
          "id",
          "org_id",
          "email",
          "role",
          "status",
          "created_at",
          "expires_at",
          "delivery_status"
        ],
        "properties": {
          "accepted_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "delivery_status": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "last_sent_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "org_id": {
            "type": "string",
            "format": "uuid"
          },
          "revoked_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "role": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "UserRow": {
        "type": "object",
        "required": [
          "id",
          "primary_email",
          "created_at"
        ],
        "properties": {
          "avatar_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "last_seen_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "primary_email": {
            "type": "string"
          }
        }
      },
      "UserSessionRow": {
        "type": "object",
        "required": [
          "id",
          "user_id",
          "org_id",
          "metadata",
          "created_at",
          "expires_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "last_seen_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "metadata": {
            "type": "object"
          },
          "org_id": {
            "type": "string",
            "format": "uuid"
          },
          "revoked_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerApiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "InstantML SDK API key sent as Authorization: Bearer instantml_..."
      },
      "bootstrapToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-InstantML-Bootstrap-Token",
        "description": "Operator-only bootstrap token for initial users, orgs, and admin key paths"
      },
      "browserSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "instantml_session",
        "description": "HttpOnly browser session cookie issued by /api/auth/dev/google or /api/auth/clerk"
      }
    }
  },
  "security": [
    {
      "bearerApiKey": []
    },
    {
      "browserSession": []
    }
  ],
...
}